FEAT: 2023.03 Dependencies Update - updated dependencies - updated CHANGELOG.md - switched from Mocha and Chai to Jest
