+++ .idea/codeStyles/Project.xml
@@ -1,0 +1,7 @@
+<component name="ProjectCodeStyleConfiguration">
+  <code_scheme name="Project" version="173">
+    <HTMLCodeStyleSettings>
+      <option name="HTML_QUOTE_STYLE" value="Single" />
+    </HTMLCodeStyleSettings>
+  </code_scheme>
+</component>
