+++ .idea/jsLinters/eslint.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="EslintConfiguration">
+    <files-pattern value="**/*.{js,ts,jsx,tsx,html}" />
     <option name="fix-on-save" value="true" />
   </component>
 </project>
