+++ .idea/prettier.xml
@@ -3,6 +3,6 @@
   <component name="PrettierConfiguration">
     <option name="myConfigurationMode" value="AUTOMATIC" />
     <option name="myRunOnSave" value="true" />
-    <option name="myFilesPattern" value="{**/*,*}.{js,ts,jsx,tsx,vue,astro,json,md}" />
+    <option name="myFilesPattern" value="{**/*,*}.{js,ts,jsx,tsx,json,md}" />
   </component>
 </project>
