--- .mocharc.js
@@ -1,5 +1,0 @@
-module.exports = {
-  require: 'ts-node/register',
-  extensions: ['ts'],
-  spec: ['test/**/*.spec.*'],
-};
