+++ package.json
@@ -37,7 +37,7 @@
     "lint:eslint": "npx eslint --fix .",
     "lint": "npm-run-all lint:prettier lint:eslint",
     "pretest": "npm-run-all lint build",
-    "test": "cross-env TS_NODE_PROJECT='./test/tsconfig.json' npx mocha",
+    "test": "npx jest",
     "start:cli": "node ./dist/cli/csv-to-xlsx.js --help",
     "start": "npm-run-all build start:cli",
     "preversion": "npm-run-all test",
@@ -60,28 +60,26 @@
     "access": "public"
   },
   "devDependencies": {
-    "@types/chai": "^4.2.15",
+    "@jest/globals": "29.7.0",
     "@types/fs-extra": "^11.0.1",
-    "@types/mocha": "^10.0.1",
     "@types/node": "^18.16.2",
-    "@typescript-eslint/eslint-plugin": "^6.1.0",
-    "@typescript-eslint/parser": "^6.1.0",
-    "chai": "^4.3.0",
-    "cross-env": "^7.0.3",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
+    "@typescript-eslint/parser": "^7.3.1",
     "eslint": "^8.39.0",
-    "eslint-config-prettier": "^8.3.0",
-    "mocha": "^10.0.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-jest": "^27.9.0",
+    "jest": "29.7.0",
     "npm-run-all": "^4.1.5",
     "pkg": "^5.1.0",
-    "prettier": "3.0.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.0.4"
+    "prettier": "3.2.5",
+    "ts-jest": "29.1.2",
+    "typescript": "^5.4.2"
   },
   "dependencies": {
     "commander": "^11.0.0",
     "csv-parse": "^5.0.4",
     "fs-extra": "^11.1.1",
-    "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.0/xlsx-0.20.0.tgz"
+    "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.1/xlsx-0.20.1.tgz"
   },
   "engines": {
     "node": ">=16"
